# api/openapi_chatgpt_operations_v016.yaml
openapi: 3.1.0
info:
  title: SPPG Vendor and Inventory Operations
  version: 0.16.0
  description: Parse vendor WhatsApp invoices, reconcile PO/receiving/invoice/reject layers, manage operational stock, and confirm vendor payment evidence. No finance ledger transaction is created automatically by these operations.
servers:
  - url: https://sppg-finance-gpt-site-production-5b7d.up.railway.app
security:
  - bearerAuth: []
paths:
...
  /v1/inventory/from-receipt:
    post:
      operationId: postSppgReceiptToInventory
      summary: Preview or post confirmed receipt into operational stock
      description: Idempotent. Creates stock movements only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptToStockRequest'
      responses:
        '200':
          description: Inventory receipt result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptToStockResponse'
  /v1/inventory/usage:
    post:
      operationId: processSppgInventoryUsage
      summary: Preview or record actual stock usage
      description: Use commit=false first. Operational stock only, no finance transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryUsageRequest'
      responses:
        '200':
          description: Usage result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryUsageResponse'
  /v1/inventory/balance:
    get:
      operationId: getSppgInventoryBalance
      summary: Read current stock balance for one item
      parameters:
        - in: query
          name: site
          required: true
          schema: {type: string, enum: [MAJA, CEMPLANG]}
        - in: query
          name: item
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Current stock balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryBalanceResponse'
  /v1/inventory/requirement-preview:
    get:
      operationId: previewSppgInventoryRequirement
      summary: Calculate purchase requirement after current stock
      description: Useful for Wikian chicken ordered for multiple production days.
      parameters:
        - in: query
          name: site
          required: true
          schema: {type: string, enum: [MAJA, CEMPLANG]}
        - in: query
          name: item
          required: true
          schema: {type: string}
        - in: query
          name: plannedQty
          required: true
          schema: {type: number, minimum: 0}
        - in: query
          name: unit
          schema: {type: string, default: kg}
      responses:
        '200':
          description: Stock-adjusted purchase requirement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryRequirementResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ReceiptToStockRequest:
      type: object
      required: [site, goods_receipt_id, commit]
      properties:
        site: {type: string, enum: [MAJA, CEMPLANG]}
        goods_receipt_id: {type: integer}
        commit: {type: boolean, default: false}
    StockMovementPreview:
      type: object
      properties:
        sourceKey: {type: string}
        itemName: {type: string}
        qty: {type: number}
        unit: {type: [string, 'null']}
        fromLocation: {type: string}
        toLocation: {type: string}
    ReceiptToStockResponse:
      type: object
      properties:
        committed: {type: boolean}
        canCommit: {type: boolean}
        goodsReceiptId: {type: integer}
        inserted: {type: integer}
        duplicates: {type: integer}
        items:
          type: array
          items:
            $ref: '#/components/schemas/StockMovementPreview'
    InventoryUsageRequest:
      type: object
      required: [site, item_name, qty, unit, commit]
      properties:
        site: {type: string, enum: [MAJA, CEMPLANG]}
        item_name: {type: string}
        qty: {type: number, exclusiveMinimum: 0}
        unit: {type: string}
        occurred_at: {type: [string, 'null'], format: date-time}
        source_ref: {type: [string, 'null']}
        commit: {type: boolean, default: false}
    InventoryUsageResponse:
      type: object
      properties:
        committed: {type: boolean}
        duplicate: {type: boolean}
        movementId: {type: [integer, 'null']}
        site: {type: string}
        itemName: {type: string}
        balanceBefore: {type: number}
        usageQty: {type: number}
        balanceAfter: {type: number}
        unit: {type: string}
        stockWarning: {type: boolean}
    InventoryBalanceResponse:
      type: object
      properties:
        site: {type: string}
        itemName: {type: string}
        balance: {type: number}
    InventoryRequirementResponse:
      type: object
      properties:
        site: {type: string}
        itemName: {type: string}
        plannedQty: {type: number}
        stockAvailable: {type: number}
        purchaseNeeded: {type: number}
        unit: {type: string}
        financeTransactionCreated: {type: boolean}